move code to layout file
