+++ game/magic/building/layout.go
@@ -1,0 +1,182 @@
+package building
+
+import (
+    "image"
+    "slices"
+    "math/rand/v2"
+)
+
+// a position within some patch of land where this building is located. the Area field
+// implicitly contains the x,y position within the overall patch of land
+type BuildingPosition struct {
+    Building Building
+    Area image.Rectangle
+}
+
+// represents a patch of land (in between roads) that can have buildings placed on it
+type Rect struct {
+    Width int
+    Height int
+    Id int
+    Buildings []BuildingPosition
+    // true if this rect should contain the fortress
+    Fortress bool
+}
+
+func (rect *Rect) Clone() *Rect {
+    newRect := &Rect{Width: rect.Width, Height: rect.Height, Id: rect.Id, Fortress: rect.Fortress}
+    newRect.Buildings = make([]BuildingPosition, len(rect.Buildings))
+    copy(newRect.Buildings, rect.Buildings)
+    return newRect
+}
+
+func (rect *Rect) Remove(building Building) {
+    rect.Buildings = slices.DeleteFunc(rect.Buildings, func(position BuildingPosition) bool {
+        return position.Building == building
+    })
+}
+
+// compute how much space is unused in this rectangle
+func (rect *Rect) EmptySpace() int {
+    total := rect.Area()
+
+    for _, building := range rect.Buildings {
+        total -= building.Area.Dx() * building.Area.Dy()
+    }
+
+    return total
+}
+
+// try to add the building to this patch of land. returns true if successful
+// each possible point the new building could be placed is tried in a random order. if the building
+// overlaps with any existing buildings then that point is skipped
+func (rect *Rect) Add(building Building, width, height int, random *rand.Rand) bool {
+
+    if rect.EmptySpace() < width * height {
+        return false
+    }
+
+    check := func(use image.Rectangle) bool {
+        for _, existing := range rect.Buildings {
+            if existing.Area.Overlaps(use) {
+                return false
+            }
+        }
+
+        return true
+    }
+
+    // fmt.Printf("Add %v (%v, %v) to %v, %v\n", building, width, height, rect.Width, rect.Height)
+
+    for _, x := range random.Perm(rect.Width - width + 1) {
+        for _, y := range random.Perm(rect.Height - height + 1) {
+            buildingRect := image.Rect(0, 0, width, height).Add(image.Pt(x, y))
+            if check(buildingRect) {
+                rect.Buildings = append(rect.Buildings, BuildingPosition{Building: building, Area: buildingRect})
+                return true
+            }
+        }
+    }
+
+    return false
+}
+
+func (rect *Rect) Area() int {
+    return rect.Width * rect.Height
+}
+
+func cloneRects(rects []*Rect) []*Rect {
+    newRects := make([]*Rect, len(rects))
+    for i, rect := range rects {
+        newRects[i] = rect.Clone()
+    }
+
+    return newRects
+}
+
+func doLayoutIterative(buildings []Building, rects []*Rect, random *rand.Rand, count *int, maxIterations int) ([]*Rect, bool) {
+    if len(buildings) == 0 {
+        return rects, true
+    }
+
+    // fmt.Printf("Trying to add %v (%v,%v)\n", building, width, height)
+
+    type State struct {
+        Buildings []Building
+        Rects []*Rect
+        Index int
+    }
+
+    var buildingsUse []Building
+    for _, building := range buildings {
+        if building == BuildingFortress {
+            rects = cloneRects(rects)
+            // find the rect that should contain the fortress and add it immediately
+            for _, rect := range rects {
+                if rect.Fortress {
+                    width, height := building.Size()
+                    if !rect.Add(building, width, height, random) {
+                        return nil, false
+                    }
+                    break
+                }
+            }
+
+        } else {
+            buildingsUse = append(buildingsUse, building)
+        }
+    }
+    buildings = buildingsUse
+
+    // sort from biggest to smallest
+    buildings = slices.SortedFunc(slices.Values(buildings), func (a, b Building) int {
+        aWidth, aHeight := a.Size()
+        bWidth, bHeight := b.Size()
+        return (bWidth * bHeight) - (aWidth * aHeight)
+    })
+
+    var stack []State
+
+    for _, i := range rand.Perm(len(rects)) {
+        stack = append(stack, State{Buildings: buildings, Rects: rects, Index: i})
+    }
+
+    for len(stack) > 0 && *count < maxIterations {
+        *count += 1
+
+        // fmt.Printf("Stack length %v count %v\n", len(stack), *count)
+
+        use := stack[len(stack) - 1]
+        stack = stack[:len(stack) - 1]
+
+        // fmt.Printf("Buildings left %v\n", len(use.Buildings))
+
+        if len(use.Buildings) == 0 {
+            return use.Rects, true
+        }
+
+        building := use.Buildings[0]
+        width, height := building.Size()
+
+        moreRects := cloneRects(use.Rects)
+
+        // the order the patches of land are tried is random
+        rect := moreRects[use.Index]
+        // fmt.Printf("Rect %v empty space %v buildings %v\n", rect.Id, rect.EmptySpace(), len(rect.Buildings))
+        if rect.Add(building, width, height, random) {
+            // fmt.Printf("Added %v (%v,%v) to rect %v\n", building, width, height, rect.Id)
+
+            for _, i := range rand.Perm(len(rects)) {
+                stack = append(stack, State{Buildings: use.Buildings[1:], Rects: moreRects, Index: i})
+            }
+
+            // fmt.Printf("Removed %v (%v,%v) from rect %v empty=%v\n", building, width, height, rect.Id, rect.EmptySpace())
+        }
+    }
+
+    return nil, false
+}
+
+func LayoutBuildings(buildings []Building, rects []*Rect, random *rand.Rand) ([]*Rect, bool) {
+    return doLayoutIterative(buildings, rects, random, new(int), 500)
+}
